+++ src/test/scala/eventstore/StreamPublisherSpec.scala
@@ -102,7 +102,7 @@ class StreamPublisherSpec extends AbstractSubscriptionActorSpec {
     "catch events that appear in between reading and subscribing" in new SubscriptionScope {
       connection expectMsg readEvents(0)
 
-      val position = 1
+      val position = 1L
       actor ! readCompleted(2, false, event0, event1)
 
       expectEvent(event0)
@@ -157,7 +157,7 @@ class StreamPublisherSpec extends AbstractSubscriptionActorSpec {
     "stop catching events that appear in between reading and subscribing if stop received" in new SubscriptionScope {
       connection expectMsg readEvents(0)
 
-      val position = 1
+      val position = 1L
       actor ! readCompleted(2, false, event0, event1)
 
       expectEvent(event0)
@@ -183,7 +183,7 @@ class StreamPublisherSpec extends AbstractSubscriptionActorSpec {
     }
 
     "continue with subscription if no events appear in between reading and subscribing" in new SubscriptionScope {
-      val position = 0
+      val position = 0L
       connection expectMsg readEvents(position)
       actor ! readCompleted(position, endOfStream = true)
 
@@ -200,7 +200,7 @@ class StreamPublisherSpec extends AbstractSubscriptionActorSpec {
 
     "continue with subscription if no events appear in between reading and subscribing and position is given" in
       new SubscriptionScope {
-        val position = 1
+        val position = 1L
         connection expectMsg readEvents(position)
 
         actor ! readCompleted(position, endOfStream = true)
@@ -216,7 +216,7 @@ class StreamPublisherSpec extends AbstractSubscriptionActorSpec {
       }
 
     "forward events while subscribed" in new SubscriptionScope {
-      val position = 0
+      val position = 0L
       connection expectMsg readEvents(position)
       actor ! readCompleted(position, endOfStream = true)
 
@@ -513,12 +513,12 @@ class StreamPublisherSpec extends AbstractSubscriptionActorSpec {
       expectMsg(OnNext(x))
     }
 
-    def newEvent(number: Int): Event = EventRecord(streamId, EventNumber.Exact(number), mock[EventData])
+    def newEvent(number: Long): Event = EventRecord(streamId, EventNumber.Exact(number), mock[EventData])
 
-    def readEvents(x: Int) =
+    def readEvents(x: Long) =
       ReadStreamEvents(streamId, EventNumber(x), readBatchSize, Forward, resolveLinkTos = resolveLinkTos)
 
-    def readCompleted(next: Int, endOfStream: Boolean, events: Event*) = ReadStreamEventsCompleted(
+    def readCompleted(next: Long, endOfStream: Boolean, events: Event*) = ReadStreamEventsCompleted(
       events = events.toList,
       nextEventNumber = EventNumber(next),
       lastEventNumber = mock[EventNumber.Exact],
@@ -527,7 +527,7 @@ class StreamPublisherSpec extends AbstractSubscriptionActorSpec {
       direction = Forward
     )
 
-    def subscribeCompleted(x: Int) = SubscribeToStreamCompleted(x.toLong, Some(EventNumber.Exact(x)))
+    def subscribeCompleted(x: Long) = SubscribeToStreamCompleted(x.toLong, Some(EventNumber.Exact(x)))
 
     override def expectTerminatedOnFailure() = {
       val failure = new ServerErrorException("test")
